tweaking CORS release implementation
