+++ backend/src/main/java/br/com/reinan/dscatalog/security/jwt/JwtUtils.java
@@ -14,8 +14,6 @@ import org.springframework.stereotype.Component;
 import java.util.Date;
 
 @Component
-@Getter
-@Setter
 public class JwtUtils {
 
     @Value("${security.jwt.secret}")
@@ -30,7 +28,7 @@ public class JwtUtils {
                 .withClaim("id", userPrincipal.getId())
                 .withClaim("roles", userPrincipal.getAuthorities().stream().map(Role::getAuthority).toList())
                 .withExpiresAt(new Date(new Date().getTime() + tokenExpiration))
-                .sign(Algorithm.HMAC256("jwtSecret"));
+                .sign(Algorithm.HMAC256(jwtSecret));
     }
 
     public String getUsernameFromJwtToken(String token) {
